+++ CHANGELOG.md
@@ -5,6 +5,45 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v1.10.0
+
+### Changed
+
+- Backup the `wp_scripts` and `wp_styles` globals before any tests are run and
+  restore it before each HTTP request to prevent side effects from tests that
+  modify these globals.
+- Mark more hooks as not being run before each test run to provide a cleaner
+  environment for tests. Filters and actions that are not run include:
+
+    - `parse_query`
+    - `parse_request`
+    - `posts_selection`
+    - `pre_get_posts`
+    - `rest_api_init`
+    - `send_headers`
+    - `template_redirect`
+    - `wp_enqueue_scripts`
+    - `wp_footer`
+    - `wp_head`
+    - `wp_print_scripts`
+    - `wp_print_styles`
+    - `wp`
+
+### Fixed
+
+- Fixed issue with `feed()` method in `Mantle\Http_Client\Response` to account for WordPress 6.5 - 6.6;
+
+## v1.9.5
+
+### Added
+
+- Added flexible caching to the HTTP client.
+- Added `is_html()` and `html()` to `Mantle\Http_Client\Response` to check if the
+  response is HTML and to return an `HTML` instance for making assertions
+  against the HTML content.
+- Added `feed()` method to `Mantle\Http_Client\Response` to parse the body
+  as a feed using SimplePie.
+
 ## v1.9.4
 
 ### Added